requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
requests>=2.32.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
duckduckgo-search>=6.3.0
fpdf2>=2.8.0
